+++ src/Components/Inputs/Composed/AutoInput.tsx
@@ -3,7 +3,11 @@ import React, {forwardRef, RefObject} from 'react'
 import classnames from 'classnames'
 
 // Components
-import {Radio, RadioRef, RadioButtonRef} from '../../Radio/index'
+import {
+  SelectGroup,
+  SelectGroupRef,
+  SelectGroupOptionRef,
+} from '../../SelectGroup/index'
 
 // Types
 import {
@@ -24,16 +28,16 @@ export interface AutoInputProps extends StandardFunctionProps {
   onChangeMode: (mode: AutoInputMode) => void
   /** Modality of radio, either "Auto" or "Custom" */
   mode: AutoInputMode
-  /** Radio color */
+  /** SelectGroup color */
   color?: ComponentColor
-  /** Controls size of Radio & Input sub-components */
+  /** Controls size of SelectGroup & Input sub-components */
   size?: ComponentSize
-  /** Pass through ref for Radio */
-  radioRef?: RefObject<RadioRef>
-  /** Pass through ref for "Auto" RadioButton */
-  radioButtonAutoRef?: RefObject<RadioButtonRef>
-  /** Pass through ref for "Custom" RadioButton */
-  radioButtonCustomRef?: RefObject<RadioButtonRef>
+  /** Pass through ref for SelectGroup */
+  radioRef?: RefObject<SelectGroupRef>
+  /** Pass through ref for "Auto" SelectGroupOption */
+  radioButtonAutoRef?: RefObject<SelectGroupOptionRef>
+  /** Pass through ref for "Custom" SelectGroupOption */
+  radioButtonCustomRef?: RefObject<SelectGroupOptionRef>
 }
 
 export type AutoInputRef = HTMLDivElement
@@ -69,13 +73,13 @@ export const AutoInput = forwardRef<AutoInputRef, AutoInputProps>(
         className={autoInputClass}
       >
         <div className="cf-auto-input--radio">
-          <Radio.Radio
+          <SelectGroup.SelectGroup
             ref={radioRef}
             shape={ButtonShape.StretchToFit}
             size={size}
             color={color}
           >
-            <Radio.Button
+            <SelectGroup.Option
               active={mode === AutoInputMode.Auto}
               id={`${id}--${AutoInputMode.Auto}`}
               testID={`${testID}--${AutoInputMode.Auto}`}
@@ -85,8 +89,8 @@ export const AutoInput = forwardRef<AutoInputRef, AutoInputProps>(
               ref={radioButtonAutoRef}
             >
               Auto
-            </Radio.Button>
-            <Radio.Button
+            </SelectGroup.Option>
+            <SelectGroup.Option
               active={mode === AutoInputMode.Custom}
               id={`${id}--${AutoInputMode.Custom}`}
               testID={`${testID}--${AutoInputMode.Custom}`}
@@ -96,8 +100,8 @@ export const AutoInput = forwardRef<AutoInputRef, AutoInputProps>(
               ref={radioButtonCustomRef}
             >
               Custom
-            </Radio.Button>
-          </Radio.Radio>
+            </SelectGroup.Option>
+          </SelectGroup.SelectGroup>
         </div>
         {mode === AutoInputMode.Custom && (
           <div