+++ src/Components/Inputs/Toggle.tsx
@@ -18,7 +18,6 @@ import './Toggle.scss'
 import {
   Omit,
   ComponentColor,
-  ComponentStatus,
   ComponentSize,
   IconFont,
   StandardFunctionProps,
@@ -51,19 +50,19 @@ export interface ToggleProps extends Omit<StandardFunctionProps, 'id'> {
   tabIndex?: number
   /** Used to group toggles together in a form */
   name?: string
-  /** Input field value to be updated with 'on X' functions */
+  /** Togle field value to be updated with 'on X' functions */
   value?: string
   /** Text to be displayed on hover tooltip */
   titleText?: string
   /** Text to be displayed on hover tooltip when radio button is disabled */
   disabledTitleText?: string
-  /** Input Component size */
+  /** Toggle Component size */
   size?: ComponentSize
-  /** Input status state */
-  status?: ComponentStatus
+  /** Prevents the user from interacting with this component */
+  disabled?: boolean
   /** Whether or not the input receives autofocus when mounted */
   autoFocus?: boolean
-  /** Pass through for container ref */
+  /** Refers to the visible element rather than the hidden input that ref refers to */
   containerRef?: RefObject<ToggleContainerRef>
   /** Controls color of toggle */
   color?: ComponentColor
@@ -85,12 +84,12 @@ export const Toggle = forwardRef<ToggleRef, ToggleProps>(
       style,
       value = '',
       color = ComponentColor.Primary,
-      status = ComponentStatus.Default,
       onBlur,
       testID = 'toggle',
       onFocus,
       checked,
       onKeyUp,
+      disabled = false,
       children,
       tabIndex,
       onChange,
@@ -115,22 +114,22 @@ export const Toggle = forwardRef<ToggleRef, ToggleProps>(
       'cf-toggle__focused': isFocused,
       'cf-toggle__checkbox': type === InputToggleType.Checkbox,
       'cf-toggle__radio': type === InputToggleType.Radio,
-      'cf-toggle__disabled': status === ComponentStatus.Disabled,
+      'cf-toggle__disabled': disabled,
       'cf-toggle__labelled': !!React.Children.count(children),
       [`${className}`]: className,
     })
 
-    const handleInputChange = (): void => {
+    const handleClick = (): void => {
+      if (disabled) {
+        return
+      }
+
       onChange(value)
     }
 
     const handleKeyUp = (e: KeyboardEvent<HTMLLabelElement>): void => {
-      if (e.key === 'Enter') {
-        handleInputChange()
-      }
-
-      if (e.key === 'Escape') {
-        e.currentTarget.blur()
+      if (e.key === ' ') {
+        handleClick()
       }
 
       if (onKeyUp) {
@@ -162,8 +161,7 @@ export const Toggle = forwardRef<ToggleRef, ToggleProps>(
       )
     }
 
-    const title =
-      status === ComponentStatus.Disabled ? disabledTitleText : titleText
+    const title = disabled ? disabledTitleText : titleText
 
     return (
       <div className={toggleClass} style={style} ref={containerRef}>
@@ -171,27 +169,29 @@ export const Toggle = forwardRef<ToggleRef, ToggleProps>(
           id={id}
           ref={ref}
           type={type}
-          checked={checked}
-          title={title}
           name={name}
+          title={title}
           value={value}
-          autoFocus={autoFocus}
-          onChange={handleInputChange}
-          onKeyPress={onKeyPress}
-          onKeyDown={onKeyDown}
           tabIndex={-1}
+          disabled={disabled}
+          readOnly={true}
+          onKeyDown={onKeyDown}
+          autoFocus={autoFocus}
           className="cf-toggle--input"
-          disabled={status === ComponentStatus.Disabled}
-          data-testid={testID}
+          onKeyPress={onKeyPress}
+          data-testid={`${testID}--input`}
+          defaultChecked={checked}
         />
         <label
-          htmlFor={id}
-          className="cf-toggle--visual-input"
           title={title}
-          tabIndex={tabIndex}
           onBlur={handleInputBlur}
+          htmlFor={id}
           onFocus={handleInputFocus}
+          onClick={handleClick}
           onKeyUp={handleKeyUp}
+          tabIndex={tabIndex}
+          className="cf-toggle--visual-input"
+          data-testid={testID}
         >
           {indicator}
         </label>